requirements.txt
ansible
awscli
python-dotenv
pytest
requests
yamllint